Move InsertInto rules into Analyzer to fix error messages.
